Introduce --analyse-from flag in the db-analyser
Note that this flag is only optional if user is running immutable-db
only. We made that choice, because when running from ChainDB, the DB
will reply its ledger state(s) from most recent ledger. This might be
especially problematic when doing profiling - user would see
obfuscated data because additional objects would be loaded into memory
(intermediate ledgers etc). Thus we've decided (at
least for now) to not make it an option when running db-analyser with
whole ChainDB.

With this flag user can select which snapshot will be chosen to
initalize the db-analyser with. User provides the slot number and if
the snapshot with a name 'slotnumber_db-analyser' exists, the db-analyser
tool will start working starting from that snapshot (instead of
starting from genesis). This ledger will be then can be used with
processAll function (see --store-ledger as an example usage).