Server2 multinode_Sim had unexpected transitions
caused by a race with withConnectionManager. So ConnectionManager simulation had weird PopScheduleOutboundError and
leaked connections problems.

Both problems sources was due to weird race conditions caused to a wrong
assumption that the connVar could be reused if in the Terminating or
TerminatedState.

We also log TerminatingSt->TerminatedSt in the right place

Authors: Marcin Szamotulski